backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13
pydantic>=2.6
pydantic-settings>=2.2
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
pyotp>=2.9
qrcode[pil]>=7.4
pandas>=2.2
openpyxl>=3.1
thefuzz[speedup]>=0.22
python-multipart>=0.0.9
pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.27
pytest-cov>=4.0
anyio>=4.0
